Move `use` functions inside `if` block to out of the block
Even if a `use` function is inside an `if` block,
it will execute regardless of the `if` condition.
A `use` functions expands into a BEGIN block,
so it will be executed first no matter where placed in the source.

This commit moves `use` functions inside `if` block to out of the block.
For the above reasons, there is no effective difference.