Use Windows symbolic link, hard link, copy by Win32 API
Perl symlink function does not work on windows, so we use Win32 API.

On Windows, the creation of a symbolic link requires a privilege.
Therefore, if the creation of the symbolic link fails, create a hard link.
However, the creation of the hard link requires the write permission
to the original existing file.
Thus, if the hard link also fails, copy the file.